[refactor] Tournamentの状態処理関数を別ファイルに分離
