+++ spa/services/tournament/WsTournamentManager.js
@@ -1,13 +1,11 @@
-import fetchApiNoBody from "../../api/fetchApiNoBody.js";
-import fetchPlayersData from "../../api/fetchPlayersData.js";
-import { renderNeonInfo } from "../../components/NeonInfo.js";
 import config from "../../config.js";
-import SPA from "../../spa.js";
-import stateManager from "../../stateManager.js";
-import WsTournamentMatchManager from "../match/WsTournamentMatchManager.js";
 import { renderTournamentBracket } from "./TournamentBracket.js";
-import { renderPlayers, renderWinnerPlayer } from "./TournamentInfo.js";
 import createTournamentData from "./createTournamentData.js";
+import {
+  ongoingStateHandler,
+  errorStateHandler,
+  finishedStateHandler,
+} from "./tournamentStateHandler.js";
 
 const wsEventHandler = {
   canRender: true,
@@ -26,34 +24,14 @@ const wsEventHandler = {
 
     switch (state) {
       case "ongoing":
-        handleOngoing(matchesData, currentRound);
+        ongoingStateHandler(matchesData, currentRound);
         break;
       case "error":
-        renderNeonInfo("ERROR", "#FF0000");
-        stateManager.state.tournamentId = null;
-        console.error("Tournament error");
+        errorStateHandler();
         break;
-      case "finished": {
-        const oldElement = document.getElementById("neon-info");
-        const wrapper = document.createElement("div");
-        wrapper.className = "d-flex justify-content-center";
-        const newElement = document.createElement("button");
-        newElement.type = "button";
-        newElement.className = "my-5 py-3 px-5 tournament-result-button";
-        newElement.textContent = "Back To Home";
-        newElement.addEventListener("click", (event) => {
-          event.preventDefault();
-          SPA.navigate("/home", null, true);
-        });
-        wrapper.appendChild(newElement);
-        oldElement.replaceWith(wrapper);
-
-        const winnerPlayerName = await fetchWinnerPlayerName(matchesData);
-        renderWinnerPlayer(winnerPlayerName);
-        stateManager.state.tournamentId = null;
-        WsTournamentManager.disconnect();
+      case "finished":
+        finishedStateHandler(matchesData);
         break;
-      }
     }
   },
   handleOpen(message) {
@@ -110,63 +88,10 @@ const WsTournamentManager = {
   setCanRender(trueOrFalse) {
     this.eventHandler.canRender = trueOrFalse;
   },
+
   forceRenderTournament() {
     this.eventHandler.renderTournament();
   },
 };
-
-async function handleOngoing(matchesData, currentRound) {
-  const participantsForRound = matchesData
-    .filter((match) => match.round === currentRound)
-    .map((match) => match.participants)[0];
-  const matchId = matchesData
-    .filter((match) => match.round === currentRound)
-    .map((match) => match.matchId)[0];
-  const playersId = participantsForRound.map((participant) => participant.id);
-
-  const playersData = await fetchPlayersData(playersId);
-  if (!playersData) {
-    return;
-  }
-  renderPlayers(playersData[0].name, playersData[1].name);
-
-  // INFO 次の試合参加者ならSTART、そうでないならWAITを描画
-  if (playersId.includes(Number(stateManager.state.userId))) {
-    renderNeonInfo("START", "#FFFFFF");
-    prepareTournamentMatch(matchId);
-  } else {
-    renderNeonInfo("WAIT...", "#0ca5bf");
-  }
-
-  function prepareTournamentMatch(matchId) {
-    try {
-      const accessToken = sessionStorage.getItem("access_token");
-      if (!(accessToken && matchId)) {
-        console.error("Error prepareTournamentMatch");
-        return;
-      }
-      WsTournamentMatchManager.connect(accessToken, matchId);
-    } catch (error) {
-      console.error(error);
-    }
-  }
-}
-
-async function fetchWinnerPlayerName(matchesData) {
-  const winnerUserId = matchesData.reduce((latestMatch, currentMatch) => {
-    return currentMatch.round > latestMatch.round ? currentMatch : latestMatch;
-  }).winnerUserId;
-  const winnerUser = await fetchApiNoBody(
-    "GET",
-    config.userService,
-    `/users?userid=${winnerUserId}`,
-  );
-
-  // UserNameの取得に失敗したらIDを返す
-  if (winnerUser.status === null || winnerUser.status >= 400) {
-    return winnerUserId;
-  }
-  return winnerUser.data.username;
-}
 
 export default WsTournamentManager;
